refactor: Move popup classes into popups package
