+++ src/main/resources/io/rpg/gui/popups/points-earned-view.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.text.Font?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="io.rpg.gui.PointsPopupController">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="io.rpg.gui.popups.PointsPopupController">
    <children>
       <Label fx:id="label" alignment="CENTER" layoutX="3.0" layoutY="157.0" prefHeight="86.0" prefWidth="592.0" text="Label" textAlignment="CENTER">
          <font>
